Separate interface from implementations
