+++ app/src/main/java/com/github/wanderwise_inc/app/data/ItineraryRepositoryTestImpl.kt
@@ -2,63 +2,16 @@ package com.github.wanderwise_inc.app.data
 
 import android.util.Log
 import com.github.wanderwise_inc.app.model.location.Itinerary
+import com.github.wanderwise_inc.app.model.location.ItineraryLabels
 import com.github.wanderwise_inc.app.model.location.Tag
+import com.google.firebase.firestore.DocumentSnapshot
+import com.google.firebase.firestore.FirebaseFirestore
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.flow
-
-/**
- * const val ITINERARY_COLLECTION_PATH: String = "itineraries"
- *
- * /** @brief repository implementation using firestore as data source */ class
- * ItineraryRepositoryFirestoreImpl : ItineraryRepository { private val LOG_DEBUG_TAG =
- * "REPOSITORY_FIRESTORE" private val db = FirebaseFirestore.getInstance() private val
- * itineraryCollection = db.collection(ITINERARY_COLLECTION_PATH) /** @return a freshly generated
- * UID for an itinerary */ private fun genItineraryUid(): String = itineraryCollection.document().id
- *
- * override fun getPublicItineraries(): Flow<List<Itinerary>> { return flow { val snapshot =
- * itineraryCollection .whereEqualTo("visible", true) .get() .addOnSuccessListener {
- * Log.d(LOG_DEBUG_TAG, "Get success") } .addOnFailureListener { e -> Log.d(LOG_DEBUG_TAG, "Get
- * failure: err=$e") } .await() val itineraryList = snapshot.map { document ->
- * document.toObject(Itinerary::class.java) } emit(itineraryList) } }
- *
- * override fun getUserItineraries(userUid: String): Flow<List<Itinerary>> { val currUserUid =
- * FirebaseAuth.getInstance().currentUser?.uid
- *
- * return flow {
- *
- * // if the current user is non-null and equal to the queried user, return all itineraries. //
- * otherwise, we only return the public ones val query = if (currUserUid != null && currUserUid ==
- * userUid) { itineraryCollection.whereEqualTo(ItineraryLabels.USER_UID, userUid) } else {
- * itineraryCollection .whereEqualTo(ItineraryLabels.USER_UID, userUid)
- * .whereEqualTo(ItineraryLabels.VISIBLE, true) }
- *
- * val snapshot = query .get() .addOnSuccessListener { Log.d(LOG_DEBUG_TAG, "Get success") }
- * .addOnFailureListener { e -> Log.d(LOG_DEBUG_TAG, "Get failure, err=$e") } .await()
- *
- * val itineraryList = snapshot.map { document -> document.toObject(Itinerary::class.java) }
- * emit(itineraryList) } }
- *
- * override fun getItinerariesWithTags(tags: List<Tag>): Flow<List<Itinerary>> { return flow { val
- * query = itineraryCollection.whereArrayContainsAny(ItineraryLabels.TAGS, tags)
- *
- * val matches = query.get().await().map { document -> document.toObject(Itinerary::class.java) }
- * Log.d(LOG_DEBUG_TAG, "got $matches") emit(matches) } }
- *
- * override fun setItinerary(itinerary: Itinerary) { if (itinerary.uid.isBlank()) { itinerary.uid =
- * genItineraryUid() } val docRef = itineraryCollection.document(itinerary.uid) docRef
- * .set(itinerary.toMap()) .addOnSuccessListener { Log.d(LOG_DEBUG_TAG, "Set success") }
- * .addOnFailureListener { e -> Log.d(LOG_DEBUG_TAG, "Set failure: err=$e") } }
- *
- * override fun updateItinerary(oldUid: String, new: Itinerary) { assert(oldUid == new.uid) val
- * docRef = itineraryCollection.document(oldUid) docRef .update(new) .addOnSuccessListener {
- * Log.d(LOG_DEBUG_TAG, "Update success") } .addOnFailureListener { e -> Log.d(LOG_DEBUG_TAG,
- * "Update failure: err=$e") } }
- *
- * override fun deleteItinerary(itinerary: Itinerary) { val docRef =
- * itineraryCollection.document(itinerary.uid) docRef .delete() .addOnSuccessListener {
- * Log.d(LOG_DEBUG_TAG, "Delete success: uid=$itinerary.uid") } .addOnFailureListener { e ->
- * Log.d(LOG_DEBUG_TAG, "Delete failure: uid=$itinerary.uid. err=$e") } } }
- */
+import kotlinx.coroutines.suspendCancellableCoroutine
+import kotlin.coroutines.resume
+import kotlin.coroutines.resumeWithException
 
 /** @brief repository used for testing viewmodel functionality */
 class ItineraryRepositoryTestImpl : ItineraryRepository {
@@ -102,5 +55,133 @@ class ItineraryRepositoryTestImpl : ItineraryRepository {
 
   override fun deleteItinerary(itinerary: Itinerary) {
     itineraries.remove(itinerary)
+  }
+}
+
+class ItineraryRepositoryImpl(private val db: FirebaseFirestore) : ItineraryRepository {
+  private val itinerariesCollection = db.collection("itineraries")
+
+  override fun getPublicItineraries(): Flow<List<Itinerary>> {
+    return flow {
+      val itineraries =
+        suspendCancellableCoroutine<List<Itinerary>> { continuation ->
+          itinerariesCollection
+            .whereEqualTo(ItineraryLabels.VISIBLE, true)
+            .get()
+            .addOnSuccessListener { snap ->
+              val documents = snap.documents
+              val iti = documents.mapNotNull { it.toObject(Itinerary::class.java) }
+              continuation.resume(iti)
+            }
+            .addOnFailureListener { exception ->
+              continuation.resumeWithException(exception)
+            }
+        }
+      emit(itineraries)
+    }
+      .catch { emit(listOf()) }
+  }
+
+  override fun getUserItineraries(userUid: String): Flow<List<Itinerary>> {
+    return flow {
+      val itineraries =
+        suspendCancellableCoroutine<List<Itinerary>> { continuation ->
+          itinerariesCollection
+            .whereEqualTo(ItineraryLabels.USER_UID, userUid)
+            .get()
+            .addOnSuccessListener { snap ->
+              val documents = snap.documents
+              val iti = documents.mapNotNull { it.toObject(Itinerary::class.java) }
+              Log.d("ItineraryRepository", "Successfully got public itineraries")
+              continuation.resume(iti)
+            }
+            .addOnFailureListener {
+              Log.d("ItineraryRepository", "Failed to get public itineraries")
+              continuation.resumeWithException(it)
+            }
+        }
+      emit(itineraries)
+    }
+      .catch { emit(listOf()) }
+  }
+
+  override fun getItinerariesWithTags(tags: List<Tag>): Flow<List<Itinerary>> {
+    return flow {
+      val itineraries =
+        suspendCancellableCoroutine<List<Itinerary>> { continuation ->
+          itinerariesCollection
+            .whereArrayContainsAny(ItineraryLabels.TAGS, tags)
+            .get()
+            .addOnSuccessListener { snap ->
+              val documents = snap.documents
+              val iti = documents.mapNotNull { it.toObject(Itinerary::class.java) }
+              Log.d("ItineraryRepository", "Successfully got public itineraries")
+              continuation.resume(iti)
+            }
+            .addOnFailureListener {
+              Log.d("ItineraryRepository", "Failed to get public itineraries")
+              continuation.resumeWithException(it)
+            }
+        }
+      emit(itineraries)
+    }
+      .catch { emit(listOf()) }
+  }
+
+  override suspend fun getItinerary(uid: String): Itinerary {
+    val document =
+      suspendCancellableCoroutine<DocumentSnapshot> { continuation ->
+        itinerariesCollection
+          .document(uid)
+          .get()
+          .addOnSuccessListener { document -> continuation.resume(document) }
+          .addOnFailureListener { exception -> continuation.resumeWithException(exception) }
+      }
+    if (document.exists()) {
+      return document.toObject(Itinerary::class.java)!!
+    } else {
+      throw Exception("Itinerary not found")
+    }
+  }
+
+  override fun setItinerary(itinerary: Itinerary) {
+    if (itinerary.uid.isBlank()) {
+      itinerary.uid = itinerariesCollection.document().id
+    }
+    val itineraryMap = itinerary.toMap()
+    itinerariesCollection
+      .document(itinerary.uid)
+      .set(itineraryMap)
+      .addOnSuccessListener { Log.d("ItineraryRepository", "Successfully set itinerary") }
+      .addOnFailureListener {
+        Log.d("ItineraryRepository", "Failed to set itinerary")
+        throw it
+      }
+  }
+
+  override fun updateItinerary(oldUid: String, new: Itinerary) {
+    if (oldUid != new.uid) {
+      throw Exception("UIDs do not match")
+    }
+    val itineraryMap = new.toMap()
+    itinerariesCollection
+      .document(oldUid)
+      .set(itineraryMap)
+      .addOnSuccessListener { Log.d("ItineraryRepository", "Successfully updated itinerary") }
+      .addOnFailureListener {
+        Log.d("ItineraryRepository", "Failed to update itinerary")
+        throw it
+      }
+  }
+
+  override fun deleteItinerary(itinerary: Itinerary) {
+    itinerariesCollection
+      .document(itinerary.uid)
+      .delete()
+      .addOnSuccessListener { Log.d("ItineraryRepository", "Successfully deleted itinerary") }
+      .addOnFailureListener {
+        Log.d("ItineraryRepository", "Failed to delete itinerary")
+        throw it
+      }
   }
 }
